each page moved in folder, some files change names
