+++ src/App.tsx
@@ -2,9 +2,9 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import './App.css'
 
 import { NavBar } from "./components/Navbar/Navbar";
-import { HomePage } from "./pages/HomePage";
-import { RandomBrewery } from "./pages/RandomBrewery";
-import { Favorites } from "./pages/Favorites";
+import { HomePage } from "./pages/HomePage/HomePage";
+import { RandomBrewery } from "./pages/RandomBrewery/RandomBrewery";
+import { Favorites } from "./pages/Favorites/Favorites";
 
 function App() {
   return (
