+++ src/components/Navbar/Navbar.css
@@ -1,32 +1,32 @@
-.navbar-container {
-    display: flex;
-    width: 100%;
-    height: 31px;
-    background-color: #e6e6e6;
-    padding-left: 10px;
-    border-bottom: 1px solid #cacaca;
-  }
-  
-  .navbarItemsRight {
-    position: absolute;
-    right: 0px;
-  }
-  
-  .navbarItem {
-    padding: 5px 20px;
-  }
-  
-  .navbarItem:hover {
-    background-color: #5e5e5e;
-  }
-  
-  a {
-    font-size: 18px;
-    font-weight: 600;
-    color: #254397;
-    text-decoration: inherit;
-  }
-  
-  a:hover {
-    color: #becfff;
-  }
+.navbarContainer {
+  display: flex;
+  width: 100%;
+  height: 31px;
+  background-color: #e6e6e6;
+  padding-left: 10px;
+  border-bottom: 1px solid #cacaca;
+}
+
+.navbarItemsRight {
+  position: absolute;
+  right: 0px;
+}
+
+.navbarItem {
+  padding: 5px 20px;
+}
+
+.navbarItem:hover {
+  background-color: #5e5e5e;
+}
+
+.navbarLink {
+  font-size: 18px;
+  font-weight: 600;
+  color: #254397;
+  text-decoration: inherit;
+}
+
+.navbarLink:hover {
+  color: #becfff;
+}
